feat: AutoscaledPool implementation (#55)
- closes #19

The current concurrency model is "launch `desired_concurrency` worker
tasks that execute the task function in a loop". Maybe we could achieve
simpler and slightly more efficient code if each execution of the task
function had its own asyncio task, similarly to the original
implementation. Follow-up tasks would be started in the `done_callback`,
or in a recurring task (both of these would be needed).

---------

Co-authored-by: Vlada Dusek <v.dusek96@gmail.com>