chore: Remove apify dependency (#793)
Apify is not really needed by crawlee. It is used only in example code
and generated code templates. By removing apify we get rid of circular
dependency between crawlee and apify.
If someone tries to create apify based project from template in crawlee
version that is not supported by any version of apify, then it will fail
to install and user will be confused.

This is done to unbloc dependency deadlock, that can happen with
incompatible releases of crawlee and apify.

---------

Co-authored-by: Vlada Dusek <v.dusek96@gmail.com>