docs: update setting up and running in cloud (#936)
- I updated the "Setting up" and "Running in cloud" pages to make them
more beginner-friendly (hopefully) and also removed mentions of Poetry.
Including Poetry was unnecessary and could be confusing. Not to mention
in Python there are X other package managers that we could hypotetically
mention as well. So I decided to stay only with Pip and Pipx for
templates. People using a package manager other than Pip will likely
know how to handle the installation on their own.
- This was reported on Slack by @honzajavorek.