chore: switch to the common self-hosted HTTPBin instance (#965)
Switches the `crawlee-python` unit tests to use the common service
account for running the HTTPBin instance (and not the personal Apify
account).

Uses the org-wide `APIFY_HTTPBIN_TOKEN` secret. It's possible that the
personal acc token will shadow the org-wide secret - let's see how it
works in the PR checks