docs: bump `typedoc-plugin`, use `@docusaurus/faster` (#1025)
Bumps the `@apify/docusaurus-typedoc-plugin-api` to use the fixes in the
latest versions. This unpacks the `Unpack`-typed arguments in the
preparing stage (until now, this was done in rendering only). This
slightly speeds up the on-page user experience, unifies the rendering in
different components, and fixes issues with reexported symbols with
`Unpack` arguments.

Includes support for `TypeAlias` (so soft-closes #1020 - although a
manual check is recommended).

Enables experimental features from `@docusaurus/faster` for faster build
times.