fix(website): use correct image source (#1115)
The image sources as they are, are fetching the images from the root,
i.e. https://crawlee.dev . Instead, we want them to fetch from crawlee
python. There is namely one image that is different and has "Javascript"
written on it ([see if you can find it](https://crawlee.dev/python)),
but it should be "Python". The rest of the images are the same between
js and py, so it doesn't matter as much.

---

I tested it with `yarn start` for the dev mode, and `npx docusaurus
build` + `npx docusaurus serve` for a production-like build and in both
cases the images are fetched correctly