chore: fix `GuessedAtParserWarning` in BeautifulSoup (#1240)
```
GuessedAtParserWarning: No parser was explicitly specified, so I'm using the best available HTML parser for this system ("lxml"). This usually isn't a problem, but if you run this code on another system, or in a different virtual environment, it may use a different parser and behave differently.
```

Use `lxml` explicitly. If this becomes an issue for someone, we can
expose the parser option.